Code Polish v1 (#2103)
* Remove Spigot platform modules and related dependencies

This commit removes all Spigot platform-specific modules, classes, and build configurations, including `AbstractSpigotPlatform`, `NbtApiInitializer`, and corresponding project configuration files (POMs). This cleanup helps streamline the project and eliminates dependencies on Spigot and NBT-API.

* Bumped Java version usage to 21.

* Remove deprecated economy-related classes and implementations.

This commit removes obsolete economy-related classes, interfaces, and implementations. The code cleanup includes `AbstractEconomy`, `Benefit`, `DepositEconomyOperation`, and economy implementations like `Economy_Vault` and `Economy_VaultUnlocked`. These were either redundant or replaced by new approaches.

* Refactor economy package structure.

Renamed and updated references for the "economyrevamp" package to "economy" for better clarity and consistency across the project. This includes file paths, imports, and related classes. No functional changes were made.

* Remove deprecated classes and methods from the codebase.

This commit eliminates several deprecated classes and methods across various packages, including ItemExpression, ShopModerator, and VaultPermissionProvider. Redundant functionality has been consolidated or replaced with newer implementations, improving maintainability and reducing legacy code bloat.

* Fix typo in method documentation in Shop.java

Corrected "attacked" to "attached" in the JavaDoc of the `checkShop` method to ensure the comment accurately describes the method's purpose. This improves code clarity and reduces potential confusion for developers.

* Update changelog.

* Update dependencies and adjust logging and API methods

Updated several dependencies, including updating commons-lang3 to 3.18.0, and adjusted dependency scopes and exclusions for better modularity. Refactored the `getPlayerLocale` method to use `QUser` instead of `UUID`. Minor logging implementation improvements were also applied.

* Switch to `CommonUtil` and refactor method calls for consistency.

Replaced usages of `StringUtils` with `CommonUtil` and updated methods for improved readability and modernized coding conventions. Bumped version in relevant `pom.xml` files to ensure compatibility, streamlined code paths using `getFirst()` replacements, and adjusted dependency scopes for better package management.

* Refactor type declarations and optimize utility methods

Replaced `EnumSet` and `EnumMap` with `HashSet` and `HashMap` for better maintainability and flexibility. Streamlined `instanceof` type casts by adding `final` for clarity and potential performance benefits. Removed deprecated or redundant methods and added logical safeguards for more robust behavior.

* Add a config to enable/disable enchants for display items, add a workaround for PacketEvents not supporting custom enchants for itemstack conversion.

* Revert distribution changes since updated changes don't seem to work when implemented according to the CodeMC docs.

+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/util/logging/container/PlayerEconomyPreCheckLog.java
@@ -3,15 +3,17 @@ package com.ghostchu.quickshop.util.logging.container;
 import com.ghostchu.quickshop.api.obj.QUser;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PlayerEconomyPreCheckLog {
 
   private static int v = 2;
   private boolean beforeTrading;
   private String player;
-  private double holding;
+  private BigDecimal holding;
 
-  public PlayerEconomyPreCheckLog(final boolean beforeTrading, final QUser player, final double holding) {
+  public PlayerEconomyPreCheckLog(final boolean beforeTrading, final QUser player, final BigDecimal holding) {
 
     this.beforeTrading = beforeTrading;
     this.player = player.serialize();
